Extract Redux mapState and mapDispatch into containers
