+++ src/app/components/Login.jsx
@@ -1,5 +1,4 @@
 import React, { Fragment } from "react";
-import { connect } from "react-redux";
 import * as mutations from '../store/mutations';
 
 export const Login = ({ authenticateUser, authenticated, createUser, requestSignUp, noAccount }) => {
@@ -28,28 +27,3 @@ export const Login = ({ authenticateUser, authenticated, createUser, requestSign
     </Fragment>
   )
 };
-
-const mapStateToProps = ({ session }) => ({
-  authenticated: session.authenticated,
-  noAccount: session.noAccount
-});
-
-const mapDispatchToProps = dispatch => ({
-  authenticateUser(e) {
-    e.preventDefault();
-    let username = e.target['username'].value;
-    let password = e.target['password'].value;
-    dispatch(mutations.requestAuthenticateUser(username, password));
-  },
-  requestSignUp() {
-    dispatch(mutations.requestSignUp(true));
-  },
-  createUser(e) {
-    e.preventDefault();
-    let username = e.target['username'].value;
-    let password = e.target['password'].value;
-    dispatch(mutations.createUser(username, password));
-  }
-});
-
-export const ConnectedLogin = connect(mapStateToProps, mapDispatchToProps)(Login);
