+++ src/app/components/Main.jsx
@@ -1,12 +1,12 @@
 import React from "react";
-import { Provider } from "react-redux";
-import { store } from "../store";
-import { ConnectedDashboard } from "./Dashboard";
-import { ConnectedNavigation } from "./Navigation";
-import { ConnectedTaskDetail } from "./TaskDetail";
-import { ConnectedLogin } from './Login';
 import { Router, Route, Redirect  } from "react-router-dom";
 import { history } from "../store/history";
+import { Provider } from "react-redux";
+import { store } from "../store";
+import { ConnectedDashboard } from "../containers/ConnectedDashboard";
+import { ConnectedNavigation } from "../containers/ConnectedNavigation";
+import { ConnectedTaskDetail } from "../containers/ConnectedTaskDetail";
+import { ConnectedLogin } from '../containers/ConnectedLogin';
 
 const RouteGuard = Component => ({ match }) => {
   console.info("Route Guard: ", match);
