+++ src/app/components/TaskList.jsx
@@ -1,6 +1,4 @@
 import React, { Fragment } from "react";
-import { connect } from "react-redux";
-import { requestTaskCreation } from "../store/mutations";
 import { Link } from "react-router-dom";
 
 export const TaskList = ({ tasks, name, id, createNewTask }) => (
@@ -17,25 +15,3 @@ export const TaskList = ({ tasks, name, id, createNewTask }) => (
   </Fragment>
 );
 
-const mapStateToProps = (state, ownProps) => {
-  let groupID = ownProps.id;
-  return {
-    name: ownProps.name,
-    id: groupID,
-    tasks: state.tasks.filter((task) => task.group === groupID),
-  };
-};
-
-const mapDispatchToProps = (dispatch, ownProps) => {
-  return {
-    createNewTask(id) {
-      // console.log("Button Click: Create new task: ", id);
-      dispatch(requestTaskCreation(id));
-    },
-  };
-};
-
-export const ConnectedTaskList = connect(
-  mapStateToProps,
-  mapDispatchToProps
-)(TaskList);
