+++ src/app/containers/ConnectedDashboard.js
@@ -1,0 +1,10 @@
+import { connect } from "react-redux";
+import { Dashboard } from '../components/Dashboard';
+
+const mapStateToProps = state => { 
+  return {
+      groups: state.groups 
+  }
+};
+
+export const ConnectedDashboard = connect(mapStateToProps)(Dashboard);
