+++ src/app/containers/ConnectedTaskList.js
@@ -1,0 +1,26 @@
+import { connect } from "react-redux";
+import { requestTaskCreation } from "../store/mutations";
+import { TaskList } from '../components/TaskList';
+
+const mapStateToProps = (state, ownProps) => {
+  let groupID = ownProps.id;
+  return {
+    name: ownProps.name,
+    id: groupID,
+    tasks: state.tasks.filter((task) => task.group === groupID),
+  };
+};
+
+const mapDispatchToProps = (dispatch, ownProps) => {
+  return {
+    createNewTask(id) {
+      // console.log("Button Click: Create new task: ", id);
+      dispatch(requestTaskCreation(id));
+    },
+  };
+};
+
+export const ConnectedTaskList = connect(
+  mapStateToProps,
+  mapDispatchToProps
+)(TaskList);
